ai-pow H1 (DoS audit): num_tiles() u64 — fix u32 tile-count overflow
num_tiles()/num_tiles_padded()/tile_index() multiplied row_tiles by
col_tiles in u32. For an in-§4.8-envelope matrix (m,n ≤ 2^24) the
product reaches 2^42 — debug: panic (crash-DoS on a crafted block);
release: silent wrap to a wrong, smaller tile count that desyncs
challenge-index ranges.

- num_tiles/num_tiles_padded/tile_index → u64; tile_coords takes u64.
  challenge_indices is now u64-native (range/return) since tile
  indices range over num_tiles.
- validate() rejects num_tiles > u32::MAX (ParamError::TooManyTiles):
  found_idx and the spot-check challenge indices are u32-addressed
  end-to-end, so a larger count is unrepresentable. validate()
  computes the count in u64, so the check itself cannot overflow.
- Adversarial test num_tiles_no_overflow_then_rejected: m,n=2^24
  ⇒ num_tiles==2^42 exact (no overflow); u32::MAX+1 boundary
  rejected; just-under not rejected by the count gate.

Real Llama mineable GEMMs (~2^20 tiles) unaffected. Full plain-proof
roundtrip + adversarial-tamper suite green.

+++ crates/ai-pow/src/fiat_shamir.rs
@@ -106,9 +106,9 @@ pub fn transcript(label: &str, parts: &[&[u8]]) -> [u8; 32] {
 /// Derive `count` distinct indices in `0..range` from `seed`. Sampling is
 /// without-replacement over a streamed XOF with rejection-and-set.
 /// Determinism: same `(seed, count, range)` always yields the same vector.
-pub fn challenge_indices(seed: &[u8; 32], count: u32, range: u32) -> Vec<u32> {
+pub fn challenge_indices(seed: &[u8; 32], count: u32, range: u64) -> Vec<u64> {
     assert!(range > 0, "range must be > 0");
-    assert!(count <= range, "count must be <= range");
+    assert!(u64::from(count) <= range, "count must be <= range");
     let mut hasher = Hasher::new_derive_key(CTX_INDICES);
     hasher.update(seed);
     hasher.update(&count.to_le_bytes());
@@ -121,7 +121,7 @@ pub fn challenge_indices(seed: &[u8; 32], count: u32, range: u32) -> Vec<u32> {
     while chosen.len() < count as usize {
         xof.fill(&mut buf);
         let r = u64::from_le_bytes(buf);
-        let idx = (r % (range as u64)) as u32;
+        let idx = r % range;
         if !taken[idx as usize] {
             taken[idx as usize] = true;
             chosen.push(idx);
